Add final CLI tools and switch to XHMM matrix wrapping type

+++ deca-core/src/main/scala/org/bdgenomics/deca/util/Phred.scala
@@ -4,8 +4,8 @@ package org.bdgenomics.deca.util
  * Created by mlinderman on 4/11/17.
  */
 object Phred {
-  def phred(prob: BigDecimal, max: Int = 99): Int = {
-    val actual = (-10.0 * math.log10(1.0 - prob.doubleValue())).toInt
+  def phred(prob: BigDecimal, max: Double = 99.0): Double = {
+    val actual = (-10.0 * math.log10(1.0 - prob.doubleValue()))
     math.min(actual, max)
   }
 }
